Added ISEC Layers and ISEC Object
Added ISECLayers which utilize the Material and Shape classes
worked on previously to define a slice of the ISEC. An ISEC
object was added to house a list of ISECLayers, which can be
added to using the ISEC.stack method. Once all layers are stacked,
the ISEC can be simulated. Simulation is currently in progress.